oom: fix potential race between verification steps
There exists an opportunity for a race between checking the eventfd for events, and checking that the cgroup exists. Specifically:

if we read the eventfd before the cgroup is destroyed, but check that the cgroup exists after its destroyed, we will miss a real oom (because the first event triggered was an oom)

To mitigate this, we need to check the eventfd after we check for the cgroup existence. If there's another event in the fd, we know the cgroup was torn down because of an OOM kill.

Signed-off-by: Peter Hunt <pehunt@redhat.com>